refactor(fjall): rename encoding->wire_key; split snapshot codec; drop dead DecodeError variant
