+++ crates/nexus-fjall/src/lib.rs
@@ -50,7 +50,7 @@
 //! `Bytes` buffer the cursor hands out — the wire-format invariant that
 //! zero-copy decoders (rkyv, flatbuffers, `#[repr(C)]` POD) rely on for
 //! sound `&T` reads. Encoding/decoding of the *key* (stream id + version)
-//! lives in [`encoding`]; the *value* layout is owned entirely by
+//! lives in [`wire_key`]; the *value* layout is owned entirely by
 //! [`nexus_store::wire`].
 //!
 //! [`GlobalSeq`]: nexus_store::store::GlobalSeq
@@ -61,12 +61,14 @@
 )]
 
 pub mod builder;
-pub mod encoding;
 pub mod error;
 mod partition;
 mod scan;
+#[cfg(feature = "snapshot")]
+pub mod snapshot;
 pub mod store;
 mod subscription_id;
+pub mod wire_key;
 
 pub use builder::FjallStoreBuilder;
 pub use error::FjallError;
