+++ crates/nexus-fjall/tests/property_tests.rs
@@ -47,7 +47,7 @@ use bytes::Bytes;
 use futures::StreamExt;
 use nexus::Version;
 use nexus_fjall::FjallStore;
-use nexus_fjall::encoding::{
+use nexus_fjall::wire_key::{
     decode_event_key, decode_event_value, decode_stream_version, encode_event_key,
     encode_event_value, encode_stream_version,
 };
